mentora-backend/requirements.txt
Django==4.2.11
setuptools>=69.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]==3.2.6
python-dotenv==1.0.1
razorpay==1.4.1
Pillow==11.1.0
whitenoise==6.12.0
django-axes==8.3.1
gunicorn>=21.2.0
